chore(yt-types.d.ts): move AuthWay type to yt-types.d.ts
