+++ packages/ui/src/shared/constants/index.ts
@@ -6,8 +6,6 @@ export const YT_OAUTH_ACCESS_TOKEN_NAME = 'yt_oauth_access_token';
 export const YT_OAUTH_REFRESH_TOKEN_NAME = 'yt_oauth_refresh_token';
 export const YT_UI_CLUSTER_HEADER_NAME = 'x-yt-ui-cluster-name';
 
-export type AuthWay = 'oauth' | 'passwd';
-
 /**
  * Allows to succeed requests when master are in read-only mode and the transaction coordinator cell is evicted
  */
